### 2024-02-21 f9fc4fe
small fix of processing_per_second
