### 2024-02-21 6da8607
fix additional ear edge cases
