### 2024-02-24 b963f70
Repurpose graph update interval to be general
- rename var skip_frames to update_delay
- update_delay is now also how the frame UI is updated -> less UI blocking
- can be changed during run time but should still not be set to 1
