### 2024-02-26 #6267 @jeffmacinnes
Hi @samhforbes - Thank you for consideration, but I respectfully will have to pass. I've stepped away from research and it's been ~5 years since I've done anything with automated computer vision analysis. This software looks really promising, and deserves a review from someone more up to date with the field than what I can offer.
