### 2024-02-28 #6425 @editorialbot
[REVIEW]: JeFaPaTo - A joint toolbox for blinking analysis and facial features extraction
**Submitting author:** @Timozen (<a href="http://orcid.org/0000-0002-6879-552X">Tim Büchner</a>)
**Repository:** https://github.com/cvjena/JeFaPaTo
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
