### 2024-03-06 cvjena/JeFaPaTo#2 @mprib
Author Contributions
As part of the JOSS review process I'm opening an issue here and will link to it from the Review Checklist. If I'm misunderstanding the workflow, someone please let me know as this is my first JOSS review.

Based on the contribution summary [here](https://github.com/cvjena/JeFaPaTo/graphs/contributors) I'm only seeing direct code contributions from @Timozen. Would it be possible to describe the contributions of the other listed authors?

Thank you!

Mac
