### 2024-03-06 #6425 @mprib
Issue raised to confirm contributions of other listed authors: https://github.com/cvjena/JeFaPaTo/issues/2
