### 2024-03-31 #6425 @mprib
@Kevin-Mattheus-Moerman,

I've opened up a couple issues and it's my understanding the primary author is traveling out of country at the moment so things are in a bit of stasis. I had trouble confirming functionality of the package when installing the binaries on Ubuntu, so I'm going to attempt to run just from cloning the repo and see how that goes.

I should be able to complete comments on Documentation and Software Paper in the coming week.

Regards,

Mac
