### 2024-04-08 cvjena/JeFaPaTo#2 @Timozen
Here are some of the contributions the other authors made.

Oliver Mothes did the original concept of the tool and the first implementations of the software. However, these comments are somehow not visible on GitHub. I've attached parts of the commit history of our internal server (which hosts a local copy of this repo). There, he is part of the commit history and contributor to the code.

Joachim Denzler supported the tool not in a coding manner but gave recommendations for algorithms, computer vision concepts, and general experimental setups for the experiments. Hence, he only indirectly contributed to the tool that needed to be visible in the commit history.

Orlando Guntinas-Lichius is the medical expert for the software, and it is his actual brainchild. He gave feedback on the usability of the tool, assessed the relevant medical extracted features, and is the test subject in the videos shown in the examples in this repository. Therefore, he could not contribute to the code of the software itself.

![image](https://github.com/cvjena/JeFaPaTo/assets/13612390/de29d12f-417b-41df-a48f-79239e1b8a92)
