### 2024-04-09 d743527
Fix missing processing_per_second attribute
