### 2024-04-09 2807606
Refactor edge cases in EAR computation
- better input handling
- do not raise exceptions but log warnings
- handle outputs interally better
- update tests accordinlyg
- depricate test for negative landmarks (which are now valid)
