### 2024-04-09 2df5665
Fix issue with auto thresholding
- check that enough values of auto thresholding are available
- set "no threhsold" case to default to "complete" blink state
