### 2024-04-09 5b32a81
Add warning if auto-threshold fails
- inform the user about possible failure of auto threshold detection
