### 2024-04-09 #6425 @Timozen
Yes, indeed ^^ I am now fully back in the office after dealing with German bureaucracy.

I have already started to investigate some of the unwanted software behavior and shipped some fixes in the `development` branch.
However, I will ping both reviewers once there is a new stable version in the main branch!

I already really highly appreciate the work put into it :)
