### 2024-04-09 0ec852e
Blink Extraction Error Handling
- disentangle parameter checks and extraction logic
- be more aggressive in error reporting
- clean the code
