+++ frontend/ui_eye_blinking_extraction.py
@@ -7,7 +7,9 @@ import structlog
 import pyqtgraph as pg
 from qtpy import QtCore, QtGui, QtWidgets
 
+from qtpy.QtWidgets import QMessageBox
 from PyQt6.QtCore import pyqtSignal
+
 
 from jefapato import blinking
 from frontend import config, jwidgets
@@ -110,7 +112,7 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
         self.scatter_l_part = self.graph.add_scatter()
         self.scatter_r_part = self.graph.add_scatter()
 
-        # UI elements 
+        # UI elements
         self.setOrientation(QtCore.Qt.Orientation.Horizontal)
 
         self.setAcceptDrops(True)
@@ -147,7 +149,7 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
 
         # second tab is the text information
         self.table_summary = jwidgets.JTableSummary()
-        
+
         self.graph_summary_visual = pg.GraphicsLayoutWidget()
         self.summary_visual = pg.ViewBox(invertY=True, lockAspect=True, enableMenu=True, enableMouse=True)
         self.summary_visual_image = pg.ImageItem()
@@ -193,7 +195,7 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
         # dont make the groupbox changeable in height
         self.box_settings.setMinimumHeight(200)
         self.set_algo = QtWidgets.QFormLayout()
-        
+
         local = QtCore.QLocale(QtCore.QLocale.Language.English, QtCore.QLocale.Country.UnitedStates) 
         doulbe_validator = QtGui.QDoubleValidator()
         doulbe_validator.setBottom(0)
@@ -243,7 +245,7 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
         self.cb_video_fps.setCurrentIndex(4)
         self.cb_video_fps.currentIndexChanged.connect(self.compute_graph_axis)
         self.set_algo.addRow("Video FPS", self.cb_video_fps)
-        
+
         box_smooth = QtWidgets.QGroupBox("Smoothing")
         box_smooth.setCheckable(True)
         self.add_handler("smooth", box_smooth)
@@ -254,14 +256,14 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
         le_smooth_size.setValidator(int_validator)
         self.add_handler("smooth_size", le_smooth_size, mapper=I2S, default=91)
         box_smooth_layout.addRow("Window Size", le_smooth_size)
-        
+
         le_smooth_poly = QtWidgets.QLineEdit()
         le_smooth_poly.setValidator(int_validator)
         self.add_handler("smooth_poly", le_smooth_poly, mapper=I2S, default=5)
         box_smooth_layout.addRow("Polynomial Degree", le_smooth_poly)
 
         self.set_algo.addRow(box_smooth)
-        
+
         # Visual Settings #
         self.box_visuals = QtWidgets.QGroupBox("Graph Control")
         self.set_visuals = QtWidgets.QFormLayout()
@@ -282,7 +284,7 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
         # Export Settings #
         self.overwrite_export = QtWidgets.QCheckBox("Overwrite Existing File")
         self.add_handler("overwrite_export", self.overwrite_export, default=True)
-    
+
         self.format_export = QtWidgets.QComboBox()
         self.format_export.addItems(["CSV", "Excel"])
         self.format_export.setCurrentIndex(0)
[…]
