### 2024-04-10 f338dcd
Add logger info for frame selection
- add information if some face frame is shown as preview
- needed for investigation of misaligned shown frame and blink
