### 2024-04-10 cb1aec8
Improve extraction handling
- fix default parameters for smoothing
- allow the graph to go to -0.05 on y range
- fix smoothing issues
- test file for 30FPS
- add best default settings as comment
