### 2024-04-10 f662c60
Add help information
- every algorithm setting parameters gets its own help button
- create some button to show recommended extraction parameters
