+++ frontend/ui_eye_blinking_extraction.py
@@ -7,7 +7,7 @@ import structlog
 import pyqtgraph as pg
 from qtpy import QtCore, QtGui, QtWidgets
 
-from qtpy.QtWidgets import QMessageBox
+from qtpy.QtWidgets import QMessageBox, QLabel
 from PyQt6.QtCore import pyqtSignal
 
 from jefapato import blinking
@@ -106,6 +106,21 @@ def sec_to_min(seconds: float) -> str:
     seconds = int(seconds % 60)
     return f"{minutes:02d}:{seconds:02d}"
 
+def create_help_button(tooltip: str, win=None) -> QtWidgets.QPushButton:
+    """
+    Create a help button with the given tooltip.
+
+    Args:
+        tooltip (str): The tooltip text to be displayed when hovering over the button.
+        win (QtWidgets.QWidget, optional): The parent widget for the help button. Defaults to None.
+
+    Returns:
+        QtWidgets.QPushButton: The created help button.
+    """
+    help_btn = QtWidgets.QPushButton(qta.icon("fa5s.question-circle"), "")
+    help_btn.setToolTip(tooltip)
+    help_btn.clicked.connect(lambda: QMessageBox.information(win, "Help", tooltip))
+    return help_btn
 
 # TODO just make this a normal widget and not a splitter
 class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
@@ -223,7 +238,7 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
         self.box_settings = QtWidgets.QGroupBox("Algorithm Settings")
         # dont make the groupbox changeable in height
         self.box_settings.setMinimumHeight(200)
-        self.set_algo = QtWidgets.QFormLayout()
+        self.set_algo = QtWidgets.QGridLayout()
 
         local = QtCore.QLocale(QtCore.QLocale.Language.English, QtCore.QLocale.Country.UnitedStates) 
         doulbe_validator = QtGui.QDoubleValidator()
@@ -236,44 +251,73 @@ class EyeBlinkingExtraction(QtWidgets.QSplitter, config.Config):
         int_validator.setBottom(0)
         int_validator.setLocale(local)
 
+        extraction_help_button = QtWidgets.QPushButton("Eye Blinking Extraction Help Description")
+        extraction_help_button.clicked.connect(lambda: QMessageBox.information(None, "Help", blinking.HELP_DESCRIPTION))
+        self.set_algo.addWidget(extraction_help_button, 0, 0, 1, 3)
+
         le_minimum_distance = QtWidgets.QLineEdit()
         le_minimum_distance.setValidator(int_validator)
         self.add_handler("min_dist", le_minimum_distance, mapper=I2S, default=50)
-        self.set_algo.addRow("Minimum Distance", le_minimum_distance)
+
+        self.set_algo.addWidget(QLabel("Minimum Distance"), 1, 0)
+        self.set_algo.addWidget(le_minimum_distance, 1, 1)
+        self.set_algo.addWidget(create_help_button("Minimum Distance: The minimum distance between two peaks in frames. Rec: 10@30FPS, 50@240FPS"), 1, 2)   
 
         le_minimum_prominence = QtWidgets.QLineEdit()
         le_minimum_prominence.setValidator(doulbe_validator)
         self.add_handler("min_prominence", le_minimum_prominence, mapper=F2S, default=0.1)
-        self.set_algo.addRow("Minimum Prominence", le_minimum_prominence)
 
+        self.set_algo.addWidget(QLabel("Minimum Prominence"), 2, 0)
+        self.set_algo.addWidget(le_minimum_prominence, 2, 1)
[…]
