### 2024-04-10 85487dd
Update message box content for auto threshold
- make the content more informative that this is not wrong but was just not possible
- also change from `warning` to `information`
