### 2024-04-10 0668648
Add parameter settings table
- include our recommended parameters for 30 and 240 fps
