### 2024-04-15 cvjena/JeFaPaTo#2 @draabe
Hi @mprib , @Timozen ,

I've came across the same question, but with @Timozen 's explanation I'm confident that the JOSS criteria for authorship are met.
