### 2024-04-15 cvjena/JeFaPaTo#10 @draabe
JOSS paper manuscript feedback
Hi @Timozen,

here are my remarks on the software paper for the first round of review. Overall, I think it's well written and mostly contains what's required. Find my detailed feedback below (row numbers in brackets).

*Major points*

- [9-11] I feel this passage needs some clarification/extension to fully explain the topic to non-domain-experts. What do you mean by "norm" with respect to facial structures, given that you highlight the significant variations of the human face in the previous sentence? How are deviations from this norm relevant in a medical context?
- [14] How do advanced computer vision techniques help to better understand blinking w.r.t. to previous approaches?
- [Summary] The transition between both paragraphs could be a bit clearer. Also, are there any general references that could be included which may help the reader to learn more about the general field of application?
- [Statement of Need] All text following the first paragraph of this section [40-67] read as if they provide a general description of the software, but do not add much to the statement of need. I'd therefore suggest to restructure this bid, either by including it in the Summary section or by creating a new section.
- [State of the field] I'm aware this point has been adressed in a previous issue, but I'm still wondering if there are actual other packages or software products (FOSS or proprietary) that professionals use to do the same task? If yes, these should be mentioned and the difference of JeFaPaTo briefly discussed. If not, this should also be mentioned.
- [42-47] This part is a bit confusing to me: Am I correct that the blinking extraction is a two-step procedure to which you refer here as the two main parts of the tool (programming interface and GUI)? It seems to me that both parts (feature extraction and blinking detection) are in fact both included in the GUI? Then what exactly do you mean by _extendable programming interface_ and _GUI_ in this passage?
- [Functionality and Usage & Platform Support] A lot of content that is presented in this two sections should be, in my opinion, part of a standalone software documentation rather than the software paper. The paper would benefit greatly in terms of structuring and conciseness if it did not contain elements such as the description of GUI components or a table containing package dependencies. However, I guess this issue should better be discussed elsewhere (please see my comments on the documentation and general remarks on the first round of review). I'd therefore suggest to return to these sections at a later stage.

*Minor poinst*
- [44] _high_ temporal resolution (?)
- [53-55] This sentence is rather lengthy and I'm not sure if grammatically correct (not a native speaker)
- [134] Extracted Medical _ly_ relevant statistics (?)

*References*
-  I'd recommend to use in-text citations where appropriate (e.g. at [30]) throughout the manuscript. I think this can be achieved by removing the brackets, i.e., using ``@soukupovaRealTimeEyeBlink2016`` instead of ``[@soukupovaRealTimeEyeBlink2016]``.
- I believe the MediaPipe paper published at CVPR should be preferred over the arxiv preprint (https://research.google/pubs/mediapipe-a-framework-for-perceiving-and-processing-reality/)
