### 2024-04-15 #6425 @draabe
Hi @Timozen & @samhforbes ,

I've completed the first round of review and would like to give a brief general summary. Congrats to the author's for creating this software, it is feature-rich and has a lean and easy-to-use GUI. Installation works fine, and it runs smoothly on my mid-end machine. The paper is well-written but could be improved slightly (https://github.com/cvjena/JeFaPaTo/issues/10).

However, I do have one major concern regarding package documentation. If I haven't missed anything, there seems to be no standalone documentation, such as a hosted documentation, a user manual or else. Instead, the documentation is somewhat scattered between the repository's ``README.md``, ``CONTRIBUTING.md`` and several paragraphs of the software paper. It should be noted that this package is focusing on it's GUI aimed at a non-technical audience instead of providing a low-level API/CLI (again, @Timozen , please correct me if I've misinterpreted that). So, extensive API documentation is not really required. I still had considerable difficulty to navigating through the graphical interfaces, understanding the basic mechanisms, and there is not really a document to get help from.

I'd personally suggest to create a standalone documentation for the package (leaving the choice of format to the authors), including a little more detailed documentation of usage, feature/output description and usage examples/walkthroughs. I think this will greatly benefit the usability and especially help the non-technical target audience adapting it into their workflows (speaking of experience). A lot of material is already available so this should leave the authors starting from scratch, and I'm happy to provide more detailed feedback in an issue if that helps.

However, I'd politely ask you @samhforbes to weigh in on this matter. I'm a bit unsure if such a standalone documentation is required by JOSS standards (or if requirements can also be satisfied from within the software paper and ``README.md``), and what best practices are documenting an API that is actually a GUI.

Thanks, @draabe
