### 2024-04-16 41dd814
Update summary
- describe non-facial norm more
- add more references for non-experts reader for possible follow up
