+++ paper/paper.bib
@@ -104,3 +104,106 @@
   pages={62-66},
   doi={10.1109/TSMC.1979.4310076}
 }
+
+@article{kwonHighspeedCameraCharacterization2013,
+  title = {High-Speed Camera Characterization of Voluntary Eye Blinking Kinematics},
+  author = {Kwon, Kyung-Ah and Shipley, Rebecca J. and Edirisinghe, Mohan and Ezra, Daniel G. and Rose, Geoff and Best, Serena M. and Cameron, Ruth E.},
+  year = {2013},
+  month = aug,
+  journal = {Journal of the Royal Society, Interface},
+  volume = {10},
+  number = {85},
+  pages = {20130227},
+  issn = {1742-5662},
+  doi = {10.1098/rsif.2013.0227},
+  langid = {english},
+  pmcid = {PMC4043155},
+  pmid = {23760297},
+}
+
+@article{vanderwerfBlinkRecoveryPatients2007,
+  title = {Blink {{Recovery}} in {{Patients}} with {{Bell}}'s {{Palsy}}: {{A Neurophysiological}} and {{Behavioral Longitudinal Study}}},
+  shorttitle = {Blink {{Recovery}} in {{Patients}} with {{Bell}}'s {{Palsy}}},
+  author = {VanderWerf, Frans and Reits, Dik and Smit, Albertine Ellen and Metselaar, Mick},
+  year = {2007},
+  month = jan,
+  journal = {Investigative Ophthalmology \& Visual Science},
+  volume = {48},
+  number = {1},
+  pages = {203--213},
+  issn = {1552-5783},
+  doi = {10.1167/iovs.06-0499},
+  urldate = {2024-04-16},
+}
+
+@article{nuuttilaDiagnosticAccuracyGlabellar2021,
+  title = {Diagnostic Accuracy of Glabellar Tap Sign for {{Parkinson}}'s Disease},
+  author = {Nuuttila, Simo and Eklund, Mikael and Joutsa, Juho and Jaakkola, Elina and M{\"a}kinen, Elina and Honkanen, Emma A. and Lindholm, Kari and Noponen, Tommi and Ihalainen, Toni and Murtom{\"a}ki, Kirsi and Nojonen, Tanja and Levo, Reeta and Mertsalmi, Tuomas and Scheperjans, Filip and Kaasinen, Valtteri},
+  year = {2021},
+  journal = {Journal of Neural Transmission},
+  volume = {128},
+  number = {11},
+  pages = {1655--1661},
+  issn = {0300-9564},
+  doi = {10.1007/s00702-021-02391-3},
+  urldate = {2024-04-16},
+}
+
+@article{vanderwerfEyelidMovementsBehavioral2003,
+  title = {Eyelid Movements: Behavioral Studies of Blinking in Humans under Different Stimulus Conditions},
+  shorttitle = {Eyelid Movements},
+  author = {VanderWerf, Frans and Brassinga, Peter and Reits, Dik and Aramideh, Majid and {Ongerboer de Visser}, Bram},
+  year = {2003},
+  month = may,
+  journal = {Journal of Neurophysiology},
+  volume = {89},
+  number = {5},
+  pages = {2784--2796},
+  issn = {0022-3077},
+  langid = {english},
+}
+
+@article{cruzSpontaneousEyeblinkActivity2011,
+  title = {Spontaneous Eyeblink Activity},
+  author = {Cruz, Antonio A. V. and Garcia, Denny M. and Pinto, Carolina T. and Cechetti, Sheila P.},
+  year = {2011},
+  month = jan,
+  journal = {The Ocular Surface},
+  volume = {9},
+  number = {1},
+  pages = {29--41},
+  issn = {1542-0124},
+  langid = {english},
[…]
