### 2024-04-16 0c71627
Update summary
- make usage of computer vision a bit clearer
