### 2024-04-16 62329e1
Update summary
- smooth the transition between paragraphs
- include references for the reader
