### 2024-04-16 80dbe33
Update statement of need
- include more existing approaches
- one for highspeed (but only every 5ms)
- two medical ones
- create new subsection for statement of need as some kind of overview
