### 2024-04-16 c7b50b2
Fix section header
-> medical -> medically
