### 2024-04-16 47a8501
Make api and gui distinction more clear
- explain what api extendablilty means
- make api usage more clear
- make gui usage more clear
