+++ paper/paper.bib
@@ -207,3 +207,37 @@
   issn = {1558-0210},
   doi = {10.1109/TNSRE.2019.2961244},
 }
+
+@article{hochreiterMachineLearningBasedDetectingEyelid2023,
+  title = {Machine-{{Learning-Based Detecting}} of {{Eyelid Closure}} and {{Smiling Using Surface Electromyography}} of {{Auricular Muscles}} in {{Patients}} with {{Postparalytic Facial Synkinesis}}: {{A Feasibility Study}}},
+  shorttitle = {Machine-{{Learning-Based Detecting}} of {{Eyelid Closure}} and {{Smiling Using Surface Electromyography}} of {{Auricular Muscles}} in {{Patients}} with {{Postparalytic Facial Synkinesis}}},
+  author = {Hochreiter, Jakob and Hoche, Eric and Janik, Luisa and Volk, Gerd Fabian and Leistritz, Lutz and Anders, Christoph and {Guntinas-Lichius}, Orlando},
+  year = {2023},
+  month = jan,
+  journal = {Diagnostics},
+  volume = {13},
+  number = {3},
+  pages = {554},
+  publisher = {Multidisciplinary Digital Publishing Institute},
+  issn = {2075-4418},
+  doi = {10.3390/diagnostics13030554},
+  urldate = {2023-03-15},
+  langid = {english},
+}
+
+@article{chenSmartphoneBasedArtificialIntelligenceAssisted2021,
+  title = {Smartphone-{{Based Artificial Intelligence-Assisted Prediction}} for {{Eyelid Measurements}}: {{Algorithm Development}} and {{Observational Validation Study}}},
+  shorttitle = {Smartphone-{{Based Artificial Intelligence-Assisted Prediction}} for {{Eyelid Measurements}}},
+  author = {Chen, Hung-Chang and Tzeng, Shin-Shi and Hsiao, Yen-Chang and Chen, Ruei-Feng and Hung, Erh-Chien and Lee, Oscar K.},
+  year = {2021},
+  month = oct,
+  journal = {JMIR mHealth and uHealth},
+  volume = {9},
+  number = {10},
+  pages = {e32444},
+  issn = {2291-5222},
+  doi = {10.2196/32444},
+  langid = {english},
+  pmcid = {PMC8538024},
+  pmid = {34538776},
+}
