### 2024-04-16 73847d7
Fix some citations styles
