### 2024-04-16 d5a2b5b
Update Lugares et al. paper
- replace arxiv with cvpr workshop paper
