+++ paper/paper.bib
@@ -8,19 +8,13 @@
   url = {https://api.semanticscholar.org/CorpusID:35923299},
 }
 
-@misc{lugaresiMediaPipeFrameworkBuilding2019,
-  title = {{{MediaPipe}}: {{A Framework}} for {{Building Perception Pipelines}}},
-  shorttitle = {{{MediaPipe}}},
-  author = {Lugaresi, Camillo and Tang, Jiuqiang and Nash, Hadon and McClanahan, Chris and Uboweja, Esha and Hays, Michael and Zhang, Fan and Chang, Chuo-Ling and Yong, Ming Guang and Lee, Juhyun and Chang, Wan-Teh and Hua, Wei and Georg, Manfred and Grundmann, Matthias},
-  year = {2019},
-  month = jun,
-  number = {arXiv:1906.08172},
-  eprint = {1906.08172},
-  primaryclass = {cs},
-  publisher = {{arXiv}},
-  doi = {10.48550/arXiv.1906.08172},
-  archiveprefix = {arxiv},
+@inproceedings{lugaresiMediaPipeFrameworkBuilding2019,
+  title = {{{MediaPipe}}: {{A}} Framework for Perceiving and Processing Reality},
+  booktitle = {Third Workshop on Computer Vision for {{AR}}/{{VR}} at {{IEEE}} Computer Vision and Pattern Recognition ({{CVPR}}) 2019},
+  author = {Lugaresi, Camillo and Tang, Jiuqiang and Nash, Hadon and McClanahan, Chris and Uboweja, Esha and Hays, Michael and Zhang, Fan and Chang, Chuo-Ling and Yong, Ming and Lee, Juhyun and Chang, Wan-Teh and Hua, Wei and Georg, Manfred and Grundmann, Matthias},
+  year = {2019}
 }
+
 
 @article{kartynnikRealtimeFacialSurface2019a,
   title = {Real-Time {{Facial Surface Geometry}} from {{Monocular Video}} on {{Mobile GPUs}}},
