### 2024-04-16 cvjena/JeFaPaTo#10 @Timozen
Thank you very much for this valuable feedback! We highly appreciated it ^^

- we added more references to give avid readers more options to inform themselves about blinking in general and the medical context
- we also made the distinction between programming API and GUI more clear in the text
- the summary transitions have been smoothed :)
- the statement of need now includes more existing approaches.

Regarding the `[Functionality and Usage & Platform Support]`, I agree with you that the balance in the paper itself often shifts between explanation and usage documentation.
We are also aware of this issue and are planning to create such external usage documentation together with the medical students who use the software already.
