### 2024-04-16 #6425 @Timozen
Thank you once again for this valuable feedback! We included the input of @draabe, and a new version is available at the main branch.

I agree that external documentation would be helpful, and we and some medical project partners have planned it.
Regarding the JOSS standards, we would also appreciate some comments from @samhforbes on how relevant this usage documentation is.
