### 2024-04-16 #6425 @samhforbes
Hi @Timozen and @draabe thanks for this discussion.
There are two considerations here. One is the community standards, so something like the [python packaging guidelines](https://packaging.python.org/en/latest/tutorials/packaging-projects/) are relevant to ensure best practice.
I think given the package is aimed at ease of use, some sort of vignette or documentation on how to follow this through with a sample dataset - so that a new user can pick up the package and run with it is reasonable to request to satisfy the _example useage_ tickbox.
In my opinion this doesn't need to be overly long, and could be something you continue to add to in the future.
Does that make sense in terms of how to proceed? The details of it are up to your discretion, but a quick guide would help ease of use.
