### 2024-04-16 #6425 @draabe
Hi @Timozen & @samhforbes ,

thanks for the quick replies - both made sense to me and clarified the matter.

In terms of community standards, I believe that project structure (packaging, installation, testing) together with repository documentation (readme, installation, contributing) are largely in place to ensure best practices and help users/contributors to navigate around the project. In terms of usage documentation, I agree with @samhforbes in that some sort of vignette/documentation/manual should accompany the package.

@Timozen, since you are planning the latter anyway, maybe an option would be to start with a minimal version of such a documentation as part of this review, which you could then extend in the future? If you want, I can also share my UX and some of the points I found a little more difficult to understand in a separate issue.
