### 2024-04-16 cvjena/JeFaPaTo#12 @Timozen
UX issues and misunderstandings!
This issue was created to collect and categorize GUI user experience problems.
