### 2024-04-16 #6425 @Timozen
To both of you, @samhforbes and @draabe, thanks a lot!

I will leverage the GitHub wiki function to create a small documentation tutorial as a first point of action. Continuous updates will be done there :)
You can find it here: https://github.com/cvjena/JeFaPaTo/wiki/Tutorial

I will create the content stepwise to include a simple workflow other people can follow and use. I would also appreciate it if you @draabe could share your UX in this issue: https://github.com/cvjena/JeFaPaTo/issues/12
That would help me a lot to understand how your interaction was ^^

Once the tutorial documentation is satisfactory, I will let you know again!
