### 2024-04-16 #6425 @mprib
Thank you to @draabe for the suggestions regarding consolidated documentation. I also found myself hunting around for information in different places and think a central guide would be a useful addition. I'm holding off for now and will check back in when @Timozen notes that the tutorial is ready for review.

One note for @Timozen  because it may relate to the tutorial or at least the installation instructions. I'm still on Windows 10 and found everything worked with ease just cloning,  creating a standard `venv` with python3.10, using `pip install requirements.txt`, then running `main.py`. This feels simpler to me than also dealing with `conda`, but I realize that is a matter of personal preference. Just wanted to point out one path that worked in a straightforward manner on a platform that is not specifically catered to in the Stable Releases.
