### 2024-04-19 eb2520b
Delete INSTALL.md
Moved into the wiki
