### 2024-04-19 cvjena/JeFaPaTo#12 @Timozen
Thanks a lot @draabe!!!

I updated the [wiki page](https://github.com/cvjena/JeFaPaTo/wiki) accordingly and included much new information.
Many of your points are valid and should be handled/noted somewhere, at least. I hope I have added everything to the wiki page.
