### 2024-04-19 #6425 @Timozen
Hey everyone, @samhforbes, @draabe and @mprib,

I took all your recommendations to heart and created a more user-friendly documentary with the wiki functionality from GitHub.
You can find it [here](https://github.com/cvjena/JeFaPaTo/wiki).

I added the following things there:
- Updated [installation guide](https://github.com/cvjena/JeFaPaTo/wiki/Installation) (moved the INSTALL.md into its page there) and added an entry using `venv` :^)
- Added two tutorials, one for [EAR score extraction](https://github.com/cvjena/JeFaPaTo/wiki/Tutorial-%E2%80%90-EAR-Score-Extraction) and one for the [blink extraction](https://github.com/cvjena/JeFaPaTo/wiki/Tutorial-%E2%80%90-Eye-Blinking-Extraction)
- I added more general information about the models and methods used inside of `JeFaPaTo` based on the ideas and feedback @draabe gave in this [issue](https://github.com/cvjena/JeFaPaTo/issues/12) here
- I added more EAR score examples and a more extended test video in the `examples` folder of the repo
- I cleaned up the `readme.md` to address the changes

Also, minor changes to the software itself:
- `JeFaPaTo` now has a slight `help menu bar` at the top, which contains information about `About,` `License,` `Documentation,` and `Acknowledgement.`
- Also, the `Eye Blinking Extraction` GUI now has more help buttons to guide the user through setting the algorithm parameters.
- Due to the issues @mprib had during his extraction, I added more error handling and increased how errors are communicated with the user. This should be, in general, more helpful to distinguish usage and coding errors.

I pushed these changes to the main branch of the repository and I would appreciate if you can check the [wiki](https://github.com/cvjena/JeFaPaTo/wiki) itself.

Currently, the only thing remaining is building the precompiled binaries for Windows and the different MacOS versions.
This still has to be done by hand by me, so I need access to these machines in our group ^^ So, it is likely these will be added in the upcoming days.
I will also try to automate the building by using GitHub automatizations, but I don't have any experience using these build tools in that manner ^^
