### 2024-04-22 10d236a
Fix missing model in binaries
- add path for models code into pyinstaller scripts
