+++ build_linux.sh
@@ -5,5 +5,5 @@ eval "$(conda shell.bash hook)"
 conda activate jefapato
 rm -r build
 rm -r dist
-pyinstaller --console --onefile --name JeFaPaTo --add-data frontend:frontend --add-data examples:examples --icon "frontend\assets\icons\icon.ico" main.py
+pyinstaller --console --onefile --name JeFaPaTo --add-data src/jefapato:jefapato --add-data frontend:frontend --add-data examples:examples --icon "frontend\assets\icons\icon.ico" main.py
 mv dist/JeFaPaTo dist/JeFaPaTo_linux
