### 2024-04-22 9700007
Fix mac build scripts
