### 2024-04-25 #6425 @Timozen
Hey @mprib, thank you for these kind words! I fully agree with you, and I think this tutorial elevates the understanding of the tool. I am sorry that it didn't exist before and you had to struggle through the UI I built 😅
