+++ paper/paper.bib
@@ -12,7 +12,8 @@
   title = {{{MediaPipe}}: {{A}} Framework for Perceiving and Processing Reality},
   booktitle = {Third Workshop on Computer Vision for {{AR}}/{{VR}} at {{IEEE}} Computer Vision and Pattern Recognition ({{CVPR}}) 2019},
   author = {Lugaresi, Camillo and Tang, Jiuqiang and Nash, Hadon and McClanahan, Chris and Uboweja, Esha and Hays, Michael and Zhang, Fan and Chang, Chuo-Ling and Yong, Ming and Lee, Juhyun and Chang, Wan-Teh and Hua, Wei and Georg, Manfred and Grundmann, Matthias},
-  year = {2019}
+  year = {2019},
+  doi = {10.48550/arXiv.1906.08172}
 }
 
 
@@ -155,6 +156,7 @@
   pages = {2784--2796},
   issn = {0022-3077},
   langid = {english},
+  doi = {10.1152/jn.00557.2002},
 }
 
 @article{cruzSpontaneousEyeblinkActivity2011,
@@ -169,6 +171,7 @@
   issn = {1542-0124},
   langid = {english},
   pmid = {21338567},
+  doi = {10.1016/s1542-0124(11)70007-6}, 
 }
 
 @article{volkInitialSeverityMotor2017,
