### 2024-05-03 #6425 @Kevin-Mattheus-Moerman
- [ ] Please edit the Zenodo archive listed license. Currently it lists a CC-BY license but this should match your actual software license.
- [ ] Please edit the Zenodo archive listed version tab, this should read `v1.1.0`
