### 2024-05-03 97f17f3
Create CITATION.cff
