### 2024-05-03 #6425 @Timozen
@Kevin-Mattheus-Moerman, thank you very much! Everything looks fine so far; only the PDF is not yet visible when visiting the corresponding JOSS page. I just assume that the server needs some minutes to process everything.
