### 2024-05-03 #6425 @Kevin-Mattheus-Moerman
@Timozen looks like the DOI is resolving now (you may need to refresh your page).
